backend/requirements.txt
asgiref==3.5.2
astroid==2.12.12
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==4.1.2
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
openapi-codec==1.3.2
platformdirs==2.5.2
pylint==2.15.4
pytz==2022.5
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
